requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0
pdfplumber==0.10.3
google-generativeai==0.3.0
